data, yaml: split yaml loading specs into src/data/files
