+++ src/data/files/wiki-info.js
@@ -1,0 +1,10 @@
+export default ({
+  documentModes: {oneDocumentTotal},
+  thingConstructors: {WikiInfo},
+}) => ({
+  title: `Process wiki info file`,
+  file: 'wiki-info.yaml',
+
+  documentMode: oneDocumentTotal,
+  documentThing: WikiInfo,
+});
